Update to split off the Armour bonus etc.
I've changed the text for the first two bullets to make it clearer that they get double Speed for regular movement and triple Speed during Assault moves if they reach contact. The existing (and original text) caused confusion where players thought you got 6x Speed in the Assault phase!

The third bullet needed updating as they are Vehicles for *all* purposes, not just for movement. For example they don't gain +1 Armour in cover, can't be transported, and aren't lumped in with Infantry when using Macro-Weapons.