Split Speed values, Cavalry and Jump Packs
In the navigation, Cavalry has been moved to Data sheet modifiers.

Jump Packs hasn't been moved as we need to decide whether to split it into two pages or add a stub page in the list of Data sheet modifiers, or use a partial to put the same content in two pages.